requirements.txt
python-dotenv
uvicorn==0.23.2
fastapi==0.104.0
pydantic==1.9.0
SQLAlchemy~=1.4.29
psycopg2-binary==2.9.5
PyYAML==6.0
python-multipart==0.0.9
kafka-python==2.0.2